refactor(user): extract auth logic into auth routs
