+++ src/routes/index.js
@@ -3,7 +3,8 @@ const express = require('express');
 const apiRouter = express.Router();
 
 const routersInits = [
-  require('./user.routes')
+  require('./user.routes'),
+  require('./auth.routes')
 ];
 
 routersInits.forEach((routerInit) => routerInit(apiRouter));
